refactor: layout component
