+++ components/layout/Banner.js
@@ -1,10 +1,10 @@
-import { useLocalStorage } from '../hooks/useLocalStorage';
+import { useLocalStorage } from '../../hooks/useLocalStorage';
 
-export function Banner() {
+export const Banner = () => {
   const [open, setOpen] = useLocalStorage('2yranniversary', true);
   if (!open) return null;
   return (
-    <div className="relative bg-temporalblue text-spaceblack text-center py-2">
+    <div className="relative bg-gradient-to-r to-blue-100 from-blue-400 text-gray-100 text-center tracking-wide py-2">
       {/* <a className=" hover:text-blue-800" href="https://docs.temporal.io/blog/funding-announcement">
       <span role="img" aria-label="Celebration" className="hidden md:inline">
         🎉
@@ -15,8 +15,11 @@ export function Banner() {
         led by Sequoia! 🎉
       </span>
     </a> */}
-      <a className=" hover:text-blue-800" href="https://temporal.io/meetup">
-        🎂 Join our 2 year anniversary meetup!
+      <a className=" hover:underline" href="https://temporal.io/meetup">
+        <span role="img" aria-label="birthday cake">
+          🎂
+        </span>{' '}
+        Join our 2 year anniversary meetup!
         {/* <span role="img" aria-label="Celebration" className="hidden md:inline">
           to ask your questions
         </span> */}
@@ -24,7 +27,7 @@ export function Banner() {
       </a>
       <button onClick={() => setOpen(false)}>
         <svg
-          className="absolute right-0 top-0 mr-2 mt-3 md:mt-2 w-5 sm:w-6 h-5 sm:h-6 text-spaceblack"
+          className="absolute right-0 top-0 mr-2 mt-3 md:mt-2 w-5 sm:w-6 h-5 sm:h-6 text-gray-100"
           xmlns="http://www.w3.org/2000/svg"
           fill="none"
           viewBox="0 0 24 24"
@@ -39,4 +42,4 @@ export function Banner() {
       </button>
     </div>
   );
-}
+};
